task4/pdfsummarizer-quizgenerator/pyproject.toml
[project]
name = "pdfsummarizer-quizgenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "openai-agents>=0.6.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
